Refactored `rustcoalescence/src/cli/simulate/*`
