+++ rustcoalescence/src/cli/simulate/dispatch/valid/mod.rs
@@ -1,0 +1,7 @@
+mod algorithm_scenario;
+mod info;
+mod launch;
+mod partitioning;
+mod rng;
+
+pub(super) use partitioning::dispatch;
